+++ lib/rom/cassandra/tasks.rb
@@ -1,6 +1,0 @@
-# encoding: utf-8
-
-require "rake"
-require "rom-cassandra"
-
-load "tasks/db.rake"