+++ spec/integration/migrate_spec.rb
@@ -1,47 +1,48 @@
 # encoding: utf-8
 
-describe "migrate" do
+describe "migrator" do
 
   # The helper function to check the db content
   def check(query)
     Cassandra.cluster.connect.execute query
   end
 
-  let(:stdout)  { StringIO.new }
-  let(:logger)  { Logger.new stdout }
-  let(:path)    { File.expand_path("../migrate", __FILE__) }
-  let(:gateway) { ROM.finalize.env.gateways[:default] }
+  let(:stdout)   { StringIO.new }
+  let(:logger)   { Logger.new stdout }
+  let(:path)     { File.expand_path("../migrate", __FILE__) }
+  let(:migrator) { ROM.env.gateways[:default].migrator }
 
   it "works" do
     ROM.setup(:cassandra, "127.0.0.1:9042")
+    ROM.finalize
 
     expect { check "SELECT * FROM logs.users;" }.to raise_error StandardError
     expect { check "SELECT * FROM logs.logs" }.to raise_error StandardError
     expect(stdout.string).to be_empty
 
-    gateway.migrate logger: logger, path: path, version: 20150825142003
+    migrator.apply logger: logger, folders: [path], target: "20150825142003"
 
     expect { check "SELECT * FROM logs.users;" }.not_to raise_error
     expect { check "SELECT * FROM logs.logs" }.to raise_error StandardError
-    expect(stdout.string).to     include "Apply migration 20150825142003"
-    expect(stdout.string).not_to include "Apply migration 20150825142024"
+    expect(stdout.string).to     include "Migration number '20150825142003' has been applied"
+    expect(stdout.string).not_to include "Migration number '20150825142024' has been applied"
 
-    gateway.migrate logger: logger, path: path
+    migrator.apply logger: logger, folders: [path]
 
     expect { check "SELECT * FROM logs.users;" }.not_to raise_error
     expect { check "SELECT * FROM logs.logs" }.not_to raise_error
-    expect(stdout.string).to     include "Apply migration 20150825142024"
-    expect(stdout.string).not_to include "Roll back"
+    expect(stdout.string).to     include "Migration number '20150825142024' has been applied"
+    expect(stdout.string).not_to include "rolled back"
 
-    gateway.migrate logger: logger, path: path, version: 0
+    migrator.rollback logger: logger, folders: [path]
 
     expect { check "SELECT * FROM logs.users;" }.to raise_error StandardError
     expect { check "SELECT * FROM logs.logs" }.to raise_error StandardError
-    expect(stdout.string).to include "Roll back migration 20150825142024"
-    expect(stdout.string).to include "Roll back migration 20150825142003"
+    expect(stdout.string).to include "Migration number '20150825142024' has been rolled back"
+    expect(stdout.string).to include "Migration number '20150825142003' has been rolled back"
   end
 
   after { check "DROP KEYSPACE IF EXISTS rom;" }
   after { check "DROP KEYSPACE IF EXISTS logs;" }
 
-end # describe migrate
+end # describe migrator